refactor : error 패키지의 이름을 enums로 변경하라
- enum끼리 모아 의미있는 패키지 이름으로 사용하기 위해 변경
