+++ src/main/java/lotto/domain/LottoResult.java
@@ -4,6 +4,7 @@ import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import lotto.enums.LottoRank;
 
 public class LottoResult {
     private final Map<LottoRank, Integer> rankCount;
